refactor: split contracts.rs and providers.rs into focused modules
Split contracts.rs (3349 lines, 58 methods) into 6 files:
- contracts/mod.rs (666): type definitions + core CRUD
- contracts/rental.rs (674): rental request lifecycle
- contracts/provisioning.rs (931): provisioning, credentials, locks
- contracts/payment.rs (529): Stripe/ICPay, releases, receipts
- contracts/extensions.rs (262): contract extensions, renewal, events
- contracts/usage.rs (374): usage events, billing, health checks

Split providers.rs (1041 lines, 21 methods) into 4 files:
- providers/mod.rs (478): type definitions + core CRUD
- providers/external.rs (232): validators, external providers, chatwoot
- providers/auto_accept.rs (190): auto-accept rules CRUD + matching
- providers/sla.rs (160): SLA config, breach detection, alerts

Also gate 4 reseller.rs methods (test-only) with #[cfg(test)].
